moving connection logic inside environment
